Test: add merge once tests for ppn test suite

+++ src/gif_generator/image/ppn_bin.cpp
@@ -17,8 +17,7 @@ namespace gif {
       const auto blue_diff = std::max(bin1.blue_channel_avg, bin2.blue_channel_avg)
                              - std::min(bin1.blue_channel_avg, bin2.blue_channel_avg);
 
-      mse += (red_diff * red_diff + green_diff * green_diff
-                                 + blue_diff * blue_diff);
+      mse += (red_diff * red_diff + green_diff * green_diff + blue_diff * blue_diff);
 
       mse *= static_cast<double>(bin1.pixel_count * bin2.pixel_count)
              / static_cast<double>(bin1.pixel_count + bin2.pixel_count);
@@ -41,15 +40,16 @@ namespace gif {
       return distance_to_next < other.distance_to_next;
     }
     void Bin::add_bin(const Bin& other) noexcept {
-      red_channel_avg = (red_channel_avg * pixel_count + other.red_channel_avg * other.pixel_count)
-                        / (pixel_count + other.pixel_count);
-      green_channel_avg
-          = (green_channel_avg * pixel_count + other.green_channel_avg * other.pixel_count)
-            / (pixel_count + other.pixel_count);
-      blue_channel_avg
-          = (blue_channel_avg * pixel_count + other.blue_channel_avg * other.pixel_count)
-            / (pixel_count + other.pixel_count);
+      red_channel_avg = (red_channel_avg * static_cast<double>(pixel_count)
+                         + other.red_channel_avg * static_cast<double>(other.pixel_count))
+                        / static_cast<double>(pixel_count + other.pixel_count);
+      green_channel_avg = (green_channel_avg * static_cast<double>(pixel_count)
+                           + other.green_channel_avg * static_cast<double>(other.pixel_count))
+                          / static_cast<double>(pixel_count + other.pixel_count);
+      blue_channel_avg = (blue_channel_avg * static_cast<double>(pixel_count)
+                          + other.blue_channel_avg * static_cast<double>(other.pixel_count))
+                         / static_cast<double>(pixel_count + other.pixel_count);
       pixel_count += other.pixel_count;
     }
-  }  // namespace ppn
+  }  // namespace image
 }  // namespace gif
